+++ R/x3p_average.R
@@ -1,6 +1,7 @@
 #' Average an x3p object
 #' 
 #' Calculate blockwise summary statistics on the surface matrix of an x3p.
+#' If the x3p object has a mask, the mode of the mask value
 #' @param x3p x3p object
 #' @param b positive integer value, block size
 #' @param f function aggregate function
@@ -21,10 +22,35 @@ x3p_average <- function(x3p, b = 10, f = mean, ...) {
   # round x and y to block size
   df$x <- df$x %/% b
   df$y <- df$y %/% b
- # browser()
-  df <- summarize(group_by(df, x, y),
-      value = f(value, ...)
+#  browser()
+  if(!is.null(x3p$mask)) {
+    df <- ungroup(add_tally(group_by(df, x, y, mask), name="mask_n"))
+
+    df <- summarize(group_by(df, x, y),
+                    value = f(value, ...),
+                    mask = mask[which.max(mask_n)] # in case of ties, take the first
     )
+    # if(!is.null(x3p$matrix.info$Mask$Annotations)) {
+    #   browser()
+    #   colors <- x3p_mask_legend(x3p)
+    #   annotations <- data.frame(
+    #     mask_low = tolower(colors),
+    #     annotation = names(colors)
+    #     )
+    #   df <- left_join(
+    #       mutate(df, mask_low=tolower(mask)),
+    #       annotations,
+    #       by="mask_low"
+    #     )
+    #   df <- select(df, -mask_low)
+    # }
+    
+  } else {
+    df <- summarize(group_by(df, x, y),
+                    value = f(value, ...)
+    )
+  }  
+  
   # scale x and y back
   df$x <- df$x * scale * b
   df$y <- df$y * scale * b